Reject workflow if latest branch commit is a merge commit
This may extend to more valid uses than intended, but for now we need
to prevent merging PRs that have had master merged into them. This scenario
will break versioning in master.

The intended workflow at balena has always been to rebase PRs, so we have
to continue to force this for now and avoid breaking merges.

Change-type: patch
Signed-off-by: Kyle Harding <kyle@balena.io>