Use merge source when creating github release notes
The `github_finalize` job was trying to use the original source of the PR
for creating the release notes. This required the source artifact to
exist at the moment of merge which is not always true as merge can
happen at any point after creating the PR. This changes the job to use
the latest source but exclude the last two commits (merge and version)
from the release notes.

Change-type: patch