Enable external contributions
The test workflow will identify whether the PR is from a fork or not
and we conditionally run either `pull_request` or `pull_request_target`
events and skip the other to avoid duplicate workflows.

Modifications to workflow files will be ignored for `pull_request_target`
events by GitHub so we just need to make sure we aren't sharing secrets
to arbitrary code.

Change-type: minor
Signed-off-by: Kyle Harding <kyle@balena.io>