Enable pseudo-tty allocation on test steps
This wraps the shell in the steps that run `npm run test`
and `poetry run pytest` in a `script -q -e -c` to allow tests that
require pseudo-tty (e.g. for applications that work with terminal colors)
to be able to operate as they would in a local dev environment.

Change-type: minor