Create local refs for orphaned commit and tag objects
For open pull requests, the versioned commit and tag
objects are created but are intentionally not referenced
anywhere in the git tree.

As such, command line tools like git describe do not return
the current versioned tag which is sometimes used in build scripts.

So before build, test, and publish jobs we can create the references
to the orphaned objects locally. These references should already
exist on merge/finalize.

Change-type: minor
Signed-off-by: Kyle Harding <kyle@balena.io>