Only provide automatic GITHUB_TOKEN to compose vars
The docker compose tests are one of the few places
where we are providing elevated secrets to untrusted user
code beyond the workflow itself.

In the case of a GitHub token appended to compose vars,
we should use the automatic GHA token if possible and avoid
leaking a privileged PAT or ephemeral app token, even when restricted
to internal PRs.

Change-type: minor
Signed-off-by: Kyle Harding <kyle@balena.io>