BREAKING: switch to official actions/create-github-app-token
This deprecates support for the token_scope input for custom
actions which wasn't used anywhere anyway.

It also deprecates support for the installation_id input as
the installation will be derived from the repo owner.

The old token action is still in use to enable auto-merge where
custom token scopes were still required. This will be removed
in the future when custom permissions are supported by the new
action.

Resolves: https://github.com/product-os/flowzone/issues/790

Change-type: major
Signed-off-by: Kyle Harding <kyle@balena.io>