Fix version bump on merging
When the github url was moved into the yaml env var declarations the
$GH_REPO env var substitution that was previously being handled by bash
is no longer being performed which means the incorrect url is being
used to attempt the update. This change switches to the gh action
interpolation for the repository which will restore the url to its
previous state.

Change-type: patch