Increase default max parallel matrix jobs to 20
According to the team at Actuated, and observed in
our job queue, setting a max parallel matrix value
DOES NOT prevent a runner from being assigned to the
job. It takes the runner but is prevented from executing
the steps if the max value is reached.

So for now we will set the default max parallel jobs
to the same global 20 concurrent max in our Actuated plan.

This will prevent jobs from taking a runner slot only
to be stopped from execution by the GH max parallel setting.

Change-type: patch
Signed-off-by: Kyle Harding <kyle@balena.io>