Build release notes from PR body
Look for a `## Release notes` section on the PR body and generate a new
output as part of the `release_notes` job. This commit updates the job
to parse renovate PRs in a different way as regular PRs. With renovate
PRs, the job will look for a `### Release Notes` in the Renovate PR body
and create a comment that the user can edit to provide the release body.

With regular PRs, the job will look for a `## Release notes` section to
use as the release notes body.

Change-type: minor