Base versioned commit on the tip of the PR head branch
Historically we have used the PR merge commit, which is the GitHub
default, to ensure that we are testing the same merges that will
be committed to the target branch.

However, we already require that branches be rebased on the
target before merging via branch protections so there is no benefit
gained from using the merge commit.

We also would like to enable use cases where we are testing against
old versions of the target branch and not include changes from the tip
of the target.

Change-type: major
Signed-off-by: Kyle Harding <kyle@balena.io>