Fallback to an invalid ref if the checkout ref is undefined
We've seen cases where the checkout ref was undefined because
a step with the output wasn't run, or because of unexpected results
in the github context json.

Without this change, the checkout would continue but quietly
fallback to the ref or sha for the event with unexpected results.

With this change, the checkout will fail so we can fix the workflow.

Change-type: patch
Signed-off-by: Kyle Harding <kyle@balena.io>