Prepare PR release message in separate job
The work on PR #1046 made the finalize github release job dependent
on the `release_notes` input. That's a regression on the correct
behavior. This PR moves the PR comment and deploy message to a separate
job that is dependent on the input, and makes the release_notes job run
on every PR, which allows the github finalize job to always run on
merge.

This is a major change because it changes the way the release_notes
input work, even though it restores the right behavior.

Change-type: major