Use --locked option to test and build rust
This ensures that the build will use the exact dependencies in
Cargo.lock and fail if Cargo.lock has not been updated.

This is a stricter option which will could potentially cause previously
working builds to fail, but it is necessary to allow for better
reproducibility of builds.

Change-type: major