Use a single Node.js version for test & publish
Projects should be opinionated and define the desired
node version in .node_version in the root of the project.
This is supported natively via the setup-node action to
and renovate to auto-detect the version to use.

Failing this, we will still fall back to testing with
recent LTS releases but only select the first compatible
version detected, in decending order (newest first).

Change-type: major
Signed-off-by: Kyle Harding <kyle@balena.io>