Add dangerous-write to octoscan ignore list
The octoscan commit that added this check was clear
that it needs some improved regex.

As it stands it will flag any writes to GITHUB_ENV
or GITHUB_OUTPUT whether they are being used
unsafely in a following step or not.

For now let's disable it to get past all the noise
until octoscan can correctly identify improper usage.

Change-type: minor
Signed-off-by: Kyle Harding <kyle@balena.io>