Restore the "sut:latest" tag for local compose tests
The last change to the labels metadata broke conditions
when there was a tag prefix or suffixand we were expecting
to test "sut:latest" via the compose file.

We were only creating local tags with the format "sut:prefix-latest"
or "sut:latest-suffix" so docker compose couldn't find the expected
"sut:latest" tag and would rebuild, possibly with the wrong targets.

Change-type: patch
Signed-off-by: Kyle Harding <kyle@balena.io>