Add check for legacy branch protection
This will allow us to detect classic branch protection
rules that would otherwise break versioning on merge.

Unfortunately it requires admin:read to check for existing
classic protection rules, but it's better than admin:write
and it's only temporary until we are confident all legacy rules
are gone and/or versioning is moved out of Flowzone.

Change-type: minor
Signed-off-by: Kyle Harding <kyle@balena.io>