Remove CARGO_REGISTRY enviroment variable
This was being passed to the `cargo_finalize` job as
`CARGO_REGISTRY_DEFAULT` but that is the wrong way to set alternate
registries. That variable is a valid configuration but it must not point
to an URL, it must point to a registry name configured in Cargo.toml. In
fact, all the registry configuration may be configured via Cargo.toml,
making the variable a redundant interface.

Change-type: patch
See: https://doc.rust-lang.org/cargo/reference/registries.html#using-an-alternate-registry