Add support for Cargo workspaces
This adds support for Cargo’s “Virtual Workspace”: https://doc.rust-lang.org/cargo/reference/workspaces.html

If `Cargo.toml` defines a `[workspace]` section, Flowzone now looks for the fields `publish`, `version` and `rust-version` in the `[workspace.package]` section.

If a `[package]` section exists, it takes precedence over the workspace and the project is assumed to be a regular Cargo crate.

Change-type: minor