Always post deploy notes even for unedited release notes
This change is to ensure we don't filter release content based
on whether a human has reviewed the release notes content or not.

If the repo has enabled the release_notes input for an environment,
that is enough to toggle the tag and release notes generation for all
merges as long as there are notable changes detected.

Change-type: minor
Signed-off-by: Kyle Harding <kyle@balena.io>